% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Launch Antibody Analysis Tool}
\usage{
launchApp(pythonPath = "~/miniconda3/bin/python", outDirPath = "~/share/igblast_new", host = "127.0.0.1", port = 3838)
}
\description{
Launches a Shiny application that provides an interface for antibody sequence analysis,
physicochemical property analysis, and post-translational modification (PTM) analysis.
}
\details{
The Shiny application includes three main tabs:

1. \strong{Sequence Analysis}: Allows users to upload a FASTA file or paste FASTA content directly
   for sequence analysis. Users can select the sequence type (Protein or DNA), species,
   and choose whether to remove duplicate sequences. The app provides options to download
   analysis results, including IgBLAST V and J gene results.

2. \strong{Physicochemical Property}: Users can select various properties to analyze,
   such as length, molecular weight, net charge, and more. The results are displayed
   in a table and can be downloaded.

3. \strong{PTM}: Provides an interface to analyze post-translational modifications.
   The results highlight specific motifs and modifications in the sequence data.

Each tab includes options to download the analysis results in CSV or TSV format.
}
\examples{
\dontrun{
  launchApp(pythonPath = "path/to/python", outDirPath = "path/to/outdir", host = "0.0.0.0", port = 4242)
}
}
\arguments{
  \item{pythonPath}{
    The path to the Python executable. If not specified, "~/miniconda3/bin/python" is used as the default.
    Please adjust the default path as necessary based on your Python installation.
  }
  \item{outDirPath}{
    The path to the OUTDIR directory. If not specified, "~/share/igblast_new" is used as the default.
    Please adjust the default path as necessary based on your file system.
  }
  \item{host}{
    The IP address that the Shiny server should listen on. If not specified, "127.0.0.1" is used as the default,
    which restricts the access to localhost. Use "0.0.0.0" to allow access from any IP address.
  }
  \item{port}{
    The port on which the Shiny application will listen. The default is 3838. Ensure that the specified port is available
    and not blocked by firewalls or other network settings.
  }
}
